Integrate InputUI scripts into prefab
