+++ Prefabs/Runtime/InputPanel.prefab
@@ -487,8 +487,8 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 0}
-        m_TargetAssemblyTypeName: ChatdollKit.Demo.Main, Assembly-CSharp
+      - m_Target: {fileID: 3916299349383938457}
+        m_TargetAssemblyTypeName: ChatdollKit.UI.InputUI, ChatdollKit
         m_MethodName: OnWakeButton
         m_Mode: 1
         m_Arguments:
@@ -634,8 +634,8 @@ MonoBehaviour:
   m_OnSubmit:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 0}
-        m_TargetAssemblyTypeName: ChatdollKit.Demo.Main, Assembly-CSharp
+      - m_Target: {fileID: 3916299349383938457}
+        m_TargetAssemblyTypeName: ChatdollKit.UI.InputUI, ChatdollKit
         m_MethodName: OnSubmitRequestInput
         m_Mode: 1
         m_Arguments:
@@ -671,6 +671,7 @@ GameObject:
   - component: {fileID: 8054922906618185391}
   - component: {fileID: 3157776774669527275}
   - component: {fileID: 6911871965741499522}
+  - component: {fileID: 3916299349383938457}
   m_Layer: 5
   m_Name: InputPanel
   m_TagString: Untagged
@@ -737,6 +738,24 @@ MonoBehaviour:
   m_FillOrigin: 0
   m_UseSpriteMesh: 0
   m_PixelsPerUnitMultiplier: 1
+--- !u!114 &3916299349383938457
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 3039352599622598743}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 1837fdad35af14fa7801dc0a4b046b78, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  dialogController: {fileID: 0}
+  requestInput: {fileID: 5656033653885551070}
+  imagePreview: {fileID: 4504484764700038898}
+  imageIcon: {fileID: 6952106151453237643}
+  imagePathPanel: {fileID: 8333490245513182035}
+  imagePathInput: {fileID: 2562600015446477067}
 --- !u!1 &3556630732168148835
 GameObject:
   m_ObjectHideFlags: 0
@@ -1102,8 +1121,8 @@ MonoBehaviour:
   m_OnSubmit:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 0}
-        m_TargetAssemblyTypeName: ChatdollKit.Demo.Main, Assembly-CSharp
+      - m_Target: {fileID: 3916299349383938457}
+        m_TargetAssemblyTypeName: ChatdollKit.UI.InputUI, ChatdollKit
         m_MethodName: OnSubmitImagePath
         m_Mode: 1
         m_Arguments:
@@ -1325,8 +1344,8 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 0}
-        m_TargetAssemblyTypeName: ChatdollKit.Demo.Main, Assembly-CSharp
+      - m_Target: {fileID: 3916299349383938457}
+        m_TargetAssemblyTypeName: ChatdollKit.UI.InputUI, ChatdollKit
         m_MethodName: OnImageButton
         m_Mode: 1
         m_Arguments:
