+++ Scripts/UI/InputUI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1837fdad35af14fa7801dc0a4b046b78
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
